I gThe functionality of the cart wasnt working all of a sudden, I backed up one commit to fix it, I need to re apply styling

+++ public/js/js-html/cart.js
@@ -6,23 +6,22 @@ $(document).ready(function () {
     let cartTotal = 0
     //! handles
     let wrapper = $('#card-wrapper')
+    
     //! GET CART
-   const requestCart = () => {
-       getItems()
-           .then(data =>{
-               $('#number').html(data.length)
-               $('#cart-amount').html(`${data.length} Items in Cart`)
-               console.log("data", data)
-               wrapper.html('')
-               data.forEach((d,i)=> {
-                   cartID.push(d.id)
-                   cartTotal += d.price
-                   let html =
-                       `  <div class="col s12 m6 l4" id="card">
+    const requestCart = () => {
+        getItems()
+            .then(data =>{
+                $('#cart-amount').html(`${data.length} Items in Cart`)
+                console.log("data", data)
+                wrapper.html('')
+                data.forEach((d,i)=> {
+                    cartID.push(d.id)
+                    cartTotal += d.price
+                    let html =
+                        `  <div class="col s12 m6 l4">
     <div class="card horizontal">
-      <div class="card-image" id="card-image">
-        <img src="../../images/Screen%20Shot%202019-12-03%20at%201.57.03%20PM.png"
-       id="image">
+      <div class="card-image">
+        <img src="https://images.unsplash.com/photo-1498842812179-c81beecf902c?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1100&q=60">
       </div>
       <div class="card-stacked">
         <div class="card-content">
@@ -38,16 +37,16 @@ id="delete-${d.id}"><i class="material-icons">clear</i></a>
       </div>
     </div>
   </div>`
-                   wrapper.append(html)
+                    wrapper.append(html)
+                    
+                })
+                addEventListener(cartID, data)
                 
-               })
-               addEventListener(cartID, data)
-
-
-           })
-   }
-   
-   //! add click events
+                
+            })
+    }
+    
+    //! add click events
     const addEventListener = (arr, cart) => {
         arr.forEach(a => {
             $(`#delete-${a}`).click(function () {
@@ -62,15 +61,14 @@ id="delete-${d.id}"><i class="material-icons">clear</i></a>
         arr.forEach(a => {
             if(a.id == id){
                 let answer = confirm(`Are you sure you want to remove this item from your cart?`)
-                if(answer) {
-                    deleteItem(id)
-                        .then(()=> {
-                            requestCart()
-                        })
-                }
-            }
+                if(answer) deleteItem(id)
+                    .then((data)=>  {
+                        console.log("data", data)
+                        getItems()
+                    })
+                    .catch(()=> console.log("ERROR"))            }
         })
-
+        
     }
